pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "flask-socketio>=5.4.1",
    "requests>=2.32.3",
    "flask-cors>=5.0.0",
    "flask-limiter>=3.9.2",
    "eventlet>=0.38.0",
    "python-telegram-bot>=20.7",
    "psutil>=6.1.0",
]
